+++ client/src/views/Card/Card.jsx
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router-dom";
 import "./card.css";
 export default function Card({ product }) {
-  const { name, brand, image } = product;
+  const { name, brand, image, supermarket } = product;
 
     const navigate = useNavigate();
     const click = () => {
@@ -15,6 +15,8 @@ export default function Card({ product }) {
           <p className="text-title">{name}</p>
             <img className="image" src={image} alt={name} key={name}></img>
             <p className="text-brand">{brand}</p>
+            <br />
+            <p className="text-brand">{supermarket}</p>
           </div>
           <button className="card-button">More info</button>
         </div>
